fix compile warnings in lib/

+++ lib/opencl_boinc.cpp
@@ -115,7 +115,7 @@ int OPENCL_DEVICE_PROP::parse(XML_PARSER& xp, const char* end_tag) {
         if (xp.parse_str("name", name, sizeof(name))) continue;
         if (xp.parse_str("vendor", vendor, sizeof(vendor))) continue;
         if (xp.parse_ulonglong("vendor_id", ull)) {
-            vendor_id = ull;
+            vendor_id = (unsigned)ull;
             continue;
         }
         if (xp.parse_int("available", n)) {
@@ -251,7 +251,7 @@ int OPENCL_DEVICE_PROP::get_opencl_driver_revision() {
         rev=0;
       }
     }
-    opencl_driver_revision=floor(rev*100+0.5);
+    opencl_driver_revision = (int)floor(rev*100+0.5);
     return 0;
 }
 
@@ -265,7 +265,7 @@ void OPENCL_DEVICE_PROP::description(char* buf, int buflen, const char* type) {
     snprintf(s2, sizeof(s2),
         "%.64s (driver version %.64s, device version %.64s, %.2fGB, %.2fGB available, %.0f GFLOPS peak)",
         name, opencl_driver_version,
-        s1, global_mem_size/GIGA,
+        s1, (double)global_mem_size/GIGA,
         opencl_available_ram/GIGA, peak_flops/1.e9
     );
